chore(tests): move tests along with the source code
